chore: clean e2e stuff
